requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-address==0.2.4
django-allauth==0.42.0
django-annoying==0.10.6
django-anymail==7.1.0
django-classy-tags==1.0.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-impersonate==1.5
django-mail-templated==2.6.5
django-richtextfield==1.6
django-rosetta==0.9.4
django-sekizai==1.1.0
django-stronghold==0.4.0
Faker==4.18.0
idna==2.9
install==1.3.1
oauthlib==3.1.0
pendulum==2.1.2
polib==1.1.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
sentry-sdk==0.14.4
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9